requirements.txt
fastapi
uvicorn[standard]
pydantic
azure-storage-blob
